requirements/base.txt
asgiref==3.4.1
crispy-bootstrap5==0.4
Django==3.2.5
djangorestframework==3.12.4
django-allauth==0.45.0
django-crispy-forms==1.12.0
django-environ==0.4.5
django-filter==2.4.0
django-sorting-bootstrap==2.7.0
dj-rest-auth==2.1.11
pytz==2021.1
sqlparse==0.4.1
Unidecode==1.2.0
